refactor(network): integrate dencun fork slot getter function into the `Network` struct
