--- src/types.rs
@@ -1,11 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Clone, Deserialize, Serialize)]
-#[serde(rename_all = "lowercase")]
-pub enum Network {
-    Mainnet,
-    Goerli,
-    Sepolia,
-    Holesky,
-    Devnet,
-}
